GUI v1.5

+++ src/IVSMathLibrary/IVSMath.cs
@@ -38,7 +38,7 @@ namespace IVSMathLibrary
             return 0;
         }
 
-        public static int Factorial(double a)
+        public static double Factorial(int a)
         {
             return 0;
         }
